third_party/WebKit/LayoutTests/http/tests/devtools/network/load-resource-for-frontend-expected.txt

Test loading resource for frontend.


Running: testNetworkResourceNonCached
Loading resource from http://localhost:8080/devtools/network/resources/resource.php
Success: 200
Logged headers: cache-control, expires, pragma
Content: "Hello world"

Running: testNetworkResourceCached
Loading resource from http://localhost:8080/devtools/network/resources/resource.php?cached=1
Success: 200
Logged headers: cache-control, expires, last-modified
Content: "Hello world"

Running: testNetworkResourceIfModifiedSince
Loading resource from http://localhost:8080/devtools/network/resources/resource.php?cached=1
Success: 304
Logged headers: 
Content: ""

Running: testResourceFromIncorrectURL
Loading resource from resource.php
Success: 404
Logged headers: 
Content: ""

Running: testResourceWithCookie
Loading resource from http://127.0.0.1:8000/devtools/network/resources/print-cookie.php
Success: 200
Logged headers: cache-control, expires, pragma
Content: "Cookie value: TestCookieValue."
